Extracted components and moved them in components folder.
